requirements.txt
click==8.1.7
Flask==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
Werkzeug==2.3.0
